refactor: extract commands into separate files
