+++ cli/src/main/kotlin/com/wire/kalium/cli/DeleteClientCommand.kt
@@ -1,0 +1,40 @@
+package com.wire.kalium.cli
+
+import com.github.ajalt.clikt.core.CliktCommand
+import com.github.ajalt.clikt.core.PrintMessage
+import com.github.ajalt.clikt.core.requireObject
+import com.github.ajalt.clikt.parameters.options.option
+import com.github.ajalt.clikt.parameters.options.prompt
+import com.wire.kalium.logic.data.client.DeleteClientParam
+import com.wire.kalium.logic.feature.UserSessionScope
+import com.wire.kalium.logic.feature.client.DeleteClientResult
+import com.wire.kalium.logic.feature.client.SelfClientsResult
+import kotlinx.coroutines.runBlocking
+
+class DeleteClientCommand : CliktCommand(name = "delete-client") {
+
+    private val userSession by requireObject<UserSessionScope>()
+    private val password: String by option(help = "Account password").prompt("password", promptSuffix = ": ", hideInput = true)
+
+    override fun run() = runBlocking {
+        val selfClientsResult = userSession.client.selfClients()
+
+        if (selfClientsResult !is SelfClientsResult.Success) {
+            throw PrintMessage("failed to retrieve self clients")
+        }
+
+        selfClientsResult.clients.forEachIndexed { index, client ->
+            echo("$index) ${client.model ?: "Unknown"}(${client.label ?: "-"}) ${client.registrationTime}")
+        }
+
+        val clientIndex = prompt("Enter client index", promptSuffix = ": ")?.toInt() ?: throw PrintMessage("Index must be an integer")
+        val deleteClientResult =
+            userSession.client.deleteClient(DeleteClientParam(password, selfClientsResult.clients[clientIndex].id))
+
+        when (deleteClientResult) {
+            is DeleteClientResult.Failure.Generic -> throw PrintMessage("Delete client failed: ${deleteClientResult.genericFailure}")
+            DeleteClientResult.Failure.InvalidCredentials -> throw PrintMessage("Invalid credentials")
+            DeleteClientResult.Success -> echo("Client successfully deleted")
+        }
+    }
+}
